+++ packages/middleware-webpack/examples/basic/src/pages/index.tsx
@@ -3,4 +3,10 @@ import ReactDOM from "react-dom";
 import App from "../App";
 import "@alifd/next/dist/next.css";
 
+import { hierarchy } from "d3-hierarchy";
+import { parse } from "mathjs";
+import { AbortController } from "abort-controller";
+import useSwr from "swr";
+
+console.log(hierarchy, parse, AbortController, useSwr);
 ReactDOM.render(<App />, document.getElementById("root"));