+++ packages/middleware-lint/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.3](https://github.com/alibaba/dawn/compare/@dawnjs/dn-middleware-lint@3.2.2...@dawnjs/dn-middleware-lint@3.2.3) (2021-11-18)
+
+### Bug Fixes
+
+- add @babel/eslint-parser to project by default ([3df8b25](https://github.com/alibaba/dawn/commit/3df8b25ab171f5fa379e08a75bbeae98783b52ff))
+
 ## [3.2.2](https://github.com/alibaba/dawn/compare/@dawnjs/dn-middleware-lint@3.2.1...@dawnjs/dn-middleware-lint@3.2.2) (2021-11-17)
 
 **Note:** Version bump only for package @dawnjs/dn-middleware-lint